+++ src/index.ts
@@ -1,7 +1,7 @@
 import { app } from "./app";
-import { criarDocente } from "./endpoints/criarDocente";
-import { mudarDocenteTurma } from "./endpoints/mudarDocenteTurma";
-import { pegarDocentes } from "./endpoints/pegarDocentes";
+import { criarDocente } from "./endpoints/Docente/criarDocente";
+import { mudarDocenteTurma } from "./endpoints/Docente/mudarDocenteTurma";
+import { pegarDocentes } from "./endpoints/Docente/pegarDocentes";
 
 
 app.post("/docentes", criarDocente);
